refactor: Move native-specific path helper out of lib.rs
